fix: pasting slide json after copying text
