+++ frontend/src/pages/PresentationEditor.vue
@@ -97,6 +97,7 @@ import {
 	lastThumbnailTime,
 	focusedSlide,
 	insertSlide,
+	getNewSlide,
 } from '@/stores/slide'
 import {
 	resetFocus,
@@ -213,9 +214,6 @@ const handleSlideShortcuts = (e) => {
 			break
 		case 'd':
 			if (isCmdOrCtrl(e)) duplicateSlide(e)
-			break
-		case 'c':
-			if (isCmdOrCtrl(e)) copySlide(e)
 			break
 	}
 }
@@ -458,35 +456,6 @@ const changeSlide = async (index, focus = true) => {
 	}
 }
 
-const getNewSlide = (toDuplicate = false, layoutId) => {
-	let layout = null
-
-	if (toDuplicate) {
-		layout = currentSlide.value
-		layout.elements = layout.elements.map((e) => ({
-			...e,
-			refId: e.refId || generateUniqueId(),
-		}))
-	} else {
-		layout = layoutResource.data?.slides?.find((l) => l.name == layoutId)
-	}
-
-	const slide = {}
-	if (layout) {
-		slide.background = layout.background
-		slide.transition = layout.transition
-		slide.transitionDuration = layout.transitionDuration
-		slide.fadeUnmatchedElements = layout.fadeUnmatchedElements
-		slide.elements = layout.elements.map((e) => ({ ...e }))
-	}
-
-	// override metadata and generate unique IDs for elements
-	slide.name = ''
-	slide.parent = presentationId.value
-
-	return slide
-}
-
 const insertDuplicateSlide = async (index, layoutId, toDuplicate) => {
 	if (toDuplicate || !index) index = slideIndex.value
 
@@ -543,18 +512,6 @@ const replaceSlide = (layoutId) => {
 	slides.value.forEach((slide, index) => {
 		slide.idx = index + 1
 	})
-}
-
-const getCopiedSlideJSON = () => {
-	const slide = getNewSlide(true)
-	return JSON.stringify(slide)
-}
-
-const copySlide = (e) => {
-	e.preventDefault()
-	const clipboardJSON = getCopiedSlideJSON()
-	e.clipboardData?.setData('application/json', clipboardJSON)
-	toast.success('Slide copied to clipboard')
 }
 
 const resetAndSave = async () => {
