+++ frontend/src/stores/copyPaste.js
@@ -1,0 +1,203 @@
+import { ref } from 'vue'
+import { toast, call } from 'frappe-ui'
+
+import { presentationId } from '@/stores/presentation'
+import { slideIndex, insertSlide, getNewSlide } from '@/stores/slide'
+import {
+	activeElements,
+	activeElementIds,
+	focusElementId,
+	addTextElement,
+	duplicateElements,
+	resetFocus,
+} from '@/stores/element'
+
+import { useTextEditor } from '@/composables/useTextEditor'
+
+import { getDocFromHTML } from '@/utils/helpers'
+import { handleUploadedMedia } from '@/utils/mediaUploads'
+
+const { activeEditor } = useTextEditor()
+
+// Copy Handlers
+
+const isCopyTriggeredByButton = ref(false)
+
+const getCopiedElementsJSON = () => JSON.stringify(activeElements.value)
+
+const getCopiedSlideJSON = () => {
+	const slide = getNewSlide(true)
+	return JSON.stringify(slide)
+}
+
+const copiedFrom = ref({})
+
+const copySlide = (e) => {
+	const clipboardJSON = getCopiedSlideJSON()
+	e.clipboardData.setData('application/json', clipboardJSON)
+	toast.success('Slide copied to clipboard')
+}
+
+const copyElements = (e) => {
+	const clipboardJSON = getCopiedElementsJSON()
+	e.clipboardData.setData('application/json', clipboardJSON)
+	copiedFrom.value = {
+		srcPresentation: presentationId.value,
+		srcSlide: slideIndex.value,
+	}
+}
+
+const handleCopy = (e) => {
+	if (isCopyTriggeredByButton.value) return
+
+	e.preventDefault()
+	const isCopyingElements = activeElementIds.value.length > 0
+	if (isCopyingElements) {
+		copyElements(e)
+	} else {
+		copySlide(e)
+	}
+}
+
+const copyToClipboard = async (text) => {
+	isCopyTriggeredByButton.value = true
+
+	if (navigator.clipboard && window.isSecureContext) {
+		await navigator.clipboard.writeText(text)
+	} else {
+		let input = document.createElement('textarea')
+		document.body.appendChild(input)
+		input.value = text
+		input.select()
+		document.execCommand('copy')
+		document.body.removeChild(input)
+	}
+
+	isCopyTriggeredByButton.value = false
+	toast.success('Copied to clipboard')
+}
+
+// Paste Handlers
+
+const handlePastedText = async (clipboardText) => {
+	await resetFocus()
+	addTextElement(clipboardText)
+}
+
+const handlePastedJSON = async (json) => {
+	const pastedArray = Array.isArray(json) ? json : []
+
+	if (
+		pastedArray[0]?.type == 'text' &&
+		focusElementId.value &&
+		focusElementId.value != pastedArray[0].id
+	) {
+		activeEditor.value.commands.insertContent(pastedArray[0].content)
+		return
+	}
+
+	const { srcPresentation, srcSlide } = copiedFrom.value
+
+	if (srcPresentation !== presentationId.value) {
+		// if pasted elements are from a different presentation
+		// add file attachments correctly to current presentation + update docnames in json
+		json = await call('slides.slides.doctype.presentation.presentation.get_updated_json', {
+			presentation: presentationId.value,
+			json: json,
+		})
+	}
+
+	duplicateElements(null, json, srcSlide)
+}
+
+const handleSvgText = (svgText) => {
+	const svgBlob = new Blob([svgText], { type: 'image/svg+xml' })
+	const svgFile = new File([svgBlob], 'pasted-image.svg', { type: 'image/svg+xml' })
+	handleUploadedMedia([{ kind: 'file', getAsFile: () => svgFile }])
+}
+
+const handlePastedSlideJSON = async (json) => {
+	const index = slideIndex.value
+
+	insertSlide(JSON.parse(json), index)
+
+	// emit('changeSlide', index + 1)
+}
+
+const isInputElement = (el) => {
+	const activeElement = document.activeElement
+	return (
+		activeElement?.tagName == 'INPUT' ||
+		activeElement?.tagName == 'TEXTAREA' ||
+		activeElement?.isContentEditable
+	)
+}
+
+const handleClipboardText = (clipboardText) => {
+	if (clipboardText?.trim().startsWith('<svg') && clipboardText?.trim().endsWith('</svg>')) {
+		handleSvgText(clipboardText)
+	} else if (clipboardText && !focusElementId.value) {
+		handlePastedText(clipboardText)
+	}
+}
+
+const handleClipboardJSON = (clipboardJSON) => {
+	const isSlideJSON = !Array.isArray(clipboardJSON) && clipboardJSON.includes('"elements"')
+	if (isSlideJSON) {
+		return handlePastedSlideJSON(clipboardJSON)
+	}
+	return handlePastedJSON(JSON.parse(clipboardJSON))
+}
+
+const dataURLToFile = (dataURL, filename) => {
+	const [meta, base64] = dataURL.split(',')
+	const mime = meta.match(/:(.*?);/)[1]
+	const binary = atob(base64)
+	const len = binary.length
+	const buffer = new Uint8Array(len)
+
+	for (let i = 0; i < len; i++) {
+		buffer[i] = binary.charCodeAt(i)
+	}
+
+	return new File([buffer], filename, {
+		type: mime,
+		lastModified: Date.now(),
+	})
+}
+
+const getImageSrcFromHTML = (clipboardTextHTML) => {
+	const doc = getDocFromHTML(clipboardTextHTML)
+	const img = doc.querySelector('img')
+
+	if (img) return img.src
+	return null
+}
+
+const handleClipboardTextHTML = (imgSrc) => {
+	const file = dataURLToFile(imgSrc, 'pasted-image.png')
+	handleUploadedMedia([{ kind: 'file', getAsFile: () => file }])
+}
+
+const handlePaste = (e) => {
+	// do not override paste event if current element is input or content editable
+	if (isInputElement()) return
+
+	e.preventDefault()
+
+	const clipboardTextHTML = e.clipboardData.getData('text/html')
+	const imgSrc = getImageSrcFromHTML(clipboardTextHTML)
+	if (clipboardTextHTML && imgSrc && imgSrc.startsWith('data:'))
+		return handleClipboardTextHTML(imgSrc)
+
+	const clipboardJSON = e.clipboardData.getData('application/json')
+	if (clipboardJSON) return handleClipboardJSON(clipboardJSON)
+
+	const clipboardText = e.clipboardData.getData('text/plain')
+	if (clipboardText) return handleClipboardText(clipboardText)
+
+	const clipboardItems = e.clipboardData.items
+	if (clipboardItems) return handleUploadedMedia(clipboardItems)
+}
+
+export { handleCopy, handlePaste, copyToClipboard }
