+++ frontend/src/utils/helpers.ts
@@ -1,5 +1,3 @@
-import { ref } from 'vue'
-import { toast } from 'frappe-ui'
 import { getAttachmentUrl } from './mediaUploads'
 
 let isClicked = false
@@ -66,26 +64,6 @@ const handleScrollBarWheelEvent = (e: WheelEvent) => {
 
 const cloneObj = (obj: any) => JSON.parse(JSON.stringify(obj))
 
-const isCopyTriggeredByButton = ref(false)
-
-const copyToClipboard = async (text: string) => {
-	isCopyTriggeredByButton.value = true
-
-	if (navigator.clipboard && window.isSecureContext) {
-		await navigator.clipboard.writeText(text)
-	} else {
-		let input = document.createElement('textarea')
-		document.body.appendChild(input)
-		input.value = text
-		input.select()
-		document.execCommand('copy')
-		document.body.removeChild(input)
-	}
-
-	isCopyTriggeredByButton.value = false
-	toast.success('Copied to clipboard')
-}
-
 const getThumbnailCardStyles = (thumbnail: string) => {
 	const thumbnailUrl = getAttachmentUrl(thumbnail)
 	return {
@@ -111,9 +89,7 @@ export {
 	setCursorPositionAtEnd,
 	handleScrollBarWheelEvent,
 	cloneObj,
-	copyToClipboard,
 	getThumbnailCardStyles,
 	getDocFromHTML,
 	isCmdOrCtrl,
-	isCopyTriggeredByButton
 }
